refactor: egctl root cmd & config commands.
