+++ cmd/egctl/main.go
@@ -9,29 +9,12 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/spf13/cobra"
-
 	"github.com/envoyproxy/gateway/internal/cmd/egctl"
 )
 
 func main() {
-	if err := rootCommand().Execute(); err != nil {
+	if err := egctl.GetRootCommand().Execute(); err != nil {
 		fmt.Fprintln(os.Stderr, err)
 		os.Exit(1)
 	}
-}
-
-func rootCommand() *cobra.Command {
-	rootCmd := &cobra.Command{
-		Use:               "egctl",
-		Long:              "A command line utility for operating Envoy Gateway",
-		SilenceUsage:      true,
-		DisableAutoGenTag: true,
-	}
-
-	rootCmd.AddCommand(egctl.NewVersionCommand())
-	rootCmd.AddCommand(egctl.NewExperimentalCommand())
-	rootCmd.AddCommand(egctl.NewConfigCommand())
-
-	return rootCmd
 }
